backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.20
Werkzeug==3.0.0
pillow==9.5.0
